System.Ext2

Description

This library exists for me to learn about ext2 and to possibly be used as part of an assignment in a Computer Science course I will be taking soon.

Please feel free to use it to learn about ext2 and play around, but don't use it for anything that you care about for now, because I don't trust myself.

Documentation

data Superblock

Constructors

Superblock

Fields

sbInodesCount :: Word32

Total number of inodes in the file system

sbBlocksCount :: Word32

Total number of blocks in the file system

sbRBlocksCount :: Word32

Number of blocks reserved for the superuser

sbFreeBlocksCount :: Word32

Number of unallocated blocks

sbFreeInodesCount :: Word32

Number of unallocated inodes

sbFirstDataBlock :: Word32

Block number of the block containing the superblock

sbLogBlockSize :: Word32

log2(block size) - 10

sbLogFragSize :: Word32

log2(fragment size) - 10

sbBlocksPerGroup :: Word32

Number of blocks in each group

sbFragsPerGroup :: Word32

Number of gragments in each group

sbInodesPerGroup :: Word32

Number of inodes in each group

sbMTime :: Word32

Last mount time

sbWTime :: Word32

Last write time

sbMntCount :: Word16

Number of mounts since last consistency check

sbMaxMntCount :: Word16

Number of allowed mounts before requiring a consistency check

sbMagic :: Word16

ext2 signature: 0xef53

sbState :: Word16

Filesystem state

sbErrors :: Word16

What to do on an error condition

sbMinorRevLevel :: Word16

Minor portion of version

sbLastCheck :: Word32

Time of last consistency check

sbCheckInterval :: Word32

Interval between forced consistency checks

sbCreatorOs :: Word32

Operating system ID

sbRevLevel :: Word32

Major portion of version

sbDefResuid :: Word16

User ID that can use reserved blocks

sbDefResgid :: Word16

Group ID that can use reserved blocks

readSuperblock :: Get Superblock

Reads the superblock information from an ext2 filesystem. Does not skip the first 1024 bytes to where the superblock lives.

See also readExtendedSuperblock

data ExtendedSuperblock

Constructors

ExtendedSuperblock

Fields

sbFirstIno :: Word32

sbInodeSize :: Word16

sbBlockGroupNumber :: Word16

sbFeatureCompat :: Word32

sbFeatureIncompat :: Word32

sbFeatureRoCompat :: Word32

sbUuid :: ByteString

sbVolumeName :: ByteString

sbLastMounted :: ByteString

sbAlgoBitmap :: Word32

sbPreallocBlocks :: Word8

sbPreallocDirBlocks :: Word8

sbUnusedAlignment :: Word16

sbJournalUuid :: ByteString

sbJournalInum :: Word32

sbJournalDev :: Word32

sbJournalLastOrphan :: Word32

readExtendedSuperblock :: Get ExtendedSuperblock

data BlockGroupDescriptorTable

Constructors

BlockGroupDescriptorTable

Fields

bgBlockBitmap :: Word32

bgInodeBitmap :: Word32

bgInodeTable :: Word32

bgFreeBlocksCount :: Word16

bgFreeInodesCount :: Word16

bgUsedDirsCount :: Word16

readBlockGroupDescriptorTable :: Get BlockGroupDescriptorTable

Reads the block group descriptor table. The last 12 ("reserved") bytes are ignored and skipped over (consumed).

data Inode

Constructors

Inode

Fields

iMode :: Word16

iUid :: Word16

iSize :: Word32

iAtime :: Word32

iCtime :: Word32

iMtime :: Word32

iDtime :: Word32

iGid :: Word16

iLinksCount :: Word16

iBlocks :: Word32

iFlags :: Word32

iOsd1 :: Word32

iBlock :: (Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32)

iGeneration :: Word32

iFileAcl :: Word32

iDirAcl :: Word32

iFaddr :: Word32

iOsd2 :: ByteString

readInode :: Get Inode

readInodeTable :: Int -> Get [Inode]

data Directory

Constructors

Directory

Fields

dInode :: Word32

dRecLen :: Word16

dNameLen :: Word8

dFileType :: Word8

dName :: ByteString

readDirectory :: Get Directory

blockGroupCount :: Superblock -> Int

Get the number of block groups within the file system.

byteFromBlock :: Superblock -> Int -> Int

Given a Superblock, and the block number, get the byte number at which it starts.

getSuperblock :: String -> IO Superblock

Open an ext2 filesystem and parse out the Superblock.

getBGDT :: String -> IO BlockGroupDescriptorTable

Open an ext2 filesystem and parse out the BlockGroupDescriptorTable.

getAllTables :: String -> IO (Superblock, ExtendedSuperblock, BlockGroupDescriptorTable)

Parses all tables out.