System.Ext2.FSChecks

Description

This module contains several checks for testing the integrity of an ext2 filesystem.

Documentation

sbMagicValid :: Superblock -> Bool

Given a superblock, ensure that its magic number is as-expected.

sbConsistency :: Superblock -> Superblock -> Bool

Given two superblocks, ensure they are consistent. We get this for free by deriving Eq.

bgdtConsistency :: BlockGroupDescriptorTable -> BlockGroupDescriptorTable -> Bool

Given two BGDTs, ensure they are consistent. We get this for free by deriving Eq.