System.Ext2

Description

This library exists for me to learn about ext2 and to possibly be used as part of an assignment in a Computer Science course I will be taking soon.

Please feel free to use it to learn about ext2 and play around, but don't use it for anything that you care about for now, because I don't trust myself.

Superblock

data Superblock

Constructors

Superblock

Fields

sbInodesCount :: Word32

Total number of inodes in the file system

sbBlocksCount :: Word32

Total number of blocks in the file system

sbRBlocksCount :: Word32

Number of blocks reserved for the superuser

sbFreeBlocksCount :: Word32

Number of unallocated blocks

sbFreeInodesCount :: Word32

Number of unallocated inodes

sbFirstDataBlock :: Word32

Block number of the block containing the superblock

sbLogBlockSize :: Word32

log2(block size) - 10

sbLogFragSize :: Word32

log2(fragment size) - 10

sbBlocksPerGroup :: Word32

Number of blocks in each group

sbFragsPerGroup :: Word32

Number of gragments in each group

sbInodesPerGroup :: Word32

Number of inodes in each group

sbMTime :: Word32

Last mount time

sbWTime :: Word32

Last write time

sbMntCount :: Word16

Number of mounts since last consistency check

sbMaxMntCount :: Word16

Number of allowed mounts before requiring a consistency check

sbMagic :: Word16

ext2 signature: 0xef53

sbState :: Word16

Filesystem state

sbErrors :: Word16

What to do on an error condition

sbMinorRevLevel :: Word16

Minor portion of version

sbLastCheck :: Word32

Time of last consistency check

sbCheckInterval :: Word32

Interval between forced consistency checks

sbCreatorOs :: Word32

Operating system ID

sbRevLevel :: Word32

Major portion of version

sbDefResuid :: Word16

User ID that can use reserved blocks

sbDefResgid :: Word16

Group ID that can use reserved blocks

Lenses

wTime :: Lens' Superblock Word32

state :: Lens' Superblock Word16

revLevel :: Lens' Superblock Word32

rBlocksCount :: Lens' Superblock Word32

mntCount :: Lens' Superblock Word16

minorRevLevel :: Lens' Superblock Word16

maxMntCount :: Lens' Superblock Word16

magic :: Lens' Superblock Word16

mTime :: Lens' Superblock Word32

logFragSize :: Lens' Superblock Word32

logBlockSize :: Lens' Superblock Word32

lastCheck :: Lens' Superblock Word32

inodesPerGroup :: Lens' Superblock Word32

inodesCount :: Lens' Superblock Word32

freeInodesCount :: Lens' Superblock Word32

freeBlocksCount :: Lens' Superblock Word32

fragsPerGroup :: Lens' Superblock Word32

firstDataBlock :: Lens' Superblock Word32

errors :: Lens' Superblock Word16

defResuid :: Lens' Superblock Word16

defResgid :: Lens' Superblock Word16

creatorOs :: Lens' Superblock Word32

checkInterval :: Lens' Superblock Word32

blocksPerGroup :: Lens' Superblock Word32

blocksCount :: Lens' Superblock Word32

Parsers

readSuperblock :: Get Superblock

Reads the superblock information from an ext2 filesystem. Does not skip the first 1024 bytes to where the superblock lives.

See also readExtendedSuperblock

ExtendedSuperblock

data ExtendedSuperblock

Constructors

ExtendedSuperblock

Fields

sbFirstIno :: Word32

sbInodeSize :: Word16

sbBlockGroupNumber :: Word16

sbFeatureCompat :: Word32

sbFeatureIncompat :: Word32

sbFeatureRoCompat :: Word32

sbUuid :: ByteString

sbVolumeName :: ByteString

sbLastMounted :: ByteString

sbAlgoBitmap :: Word32

sbPreallocBlocks :: Word8

sbPreallocDirBlocks :: Word8

sbUnusedAlignment :: Word16

sbJournalUuid :: ByteString

sbJournalInum :: Word32

sbJournalDev :: Word32

sbJournalLastOrphan :: Word32

Lenses

volumeName :: Lens' ExtendedSuperblock ByteString

uuid :: Lens' ExtendedSuperblock ByteString

unusedAlignment :: Lens' ExtendedSuperblock Word16

preallocDirBlocks :: Lens' ExtendedSuperblock Word8

preallocBlocks :: Lens' ExtendedSuperblock Word8

lastMounted :: Lens' ExtendedSuperblock ByteString

journalUuid :: Lens' ExtendedSuperblock ByteString

journalLastOrphan :: Lens' ExtendedSuperblock Word32

journalInum :: Lens' ExtendedSuperblock Word32

journalDev :: Lens' ExtendedSuperblock Word32

inodeSize :: Lens' ExtendedSuperblock Word16

firstIno :: Lens' ExtendedSuperblock Word32

featureRoCompat :: Lens' ExtendedSuperblock Word32

featureIncompat :: Lens' ExtendedSuperblock Word32

featureCompat :: Lens' ExtendedSuperblock Word32

blockGroupNumber :: Lens' ExtendedSuperblock Word16

algoBitmap :: Lens' ExtendedSuperblock Word32

Parsers

readExtendedSuperblock :: Get ExtendedSuperblock

BlockGroupDescriptorTable

data BlockGroupDescriptorTable

Constructors

BlockGroupDescriptorTable

Fields

bgBlockBitmap :: Word32

bgInodeBitmap :: Word32

bgInodeTable :: Word32

bgFreeBlocksCountBg :: Word16

bgFreeInodesCountBg :: Word16

bgUsedDirsCount :: Word16

Lenses

usedDirsCount :: Lens' BlockGroupDescriptorTable Word16

inodeTable :: Lens' BlockGroupDescriptorTable Word32

inodeBitmap :: Lens' BlockGroupDescriptorTable Word32

freeInodesCountBg :: Lens' BlockGroupDescriptorTable Word16

freeBlocksCountBg :: Lens' BlockGroupDescriptorTable Word16

blockBitmap :: Lens' BlockGroupDescriptorTable Word32

Parsers

readBlockGroupDescriptorTable :: Get BlockGroupDescriptorTable

Reads the block group descriptor table. The last 12 ("reserved") bytes are ignored and skipped over (consumed).

Inode

data Inode

Constructors

Inode

Fields

iMode :: Word16

iUid :: Word16

iSize :: Word32

iAtime :: Word32

iCtime :: Word32

iMtime :: Word32

iDtime :: Word32

iGid :: Word16

iLinksCount :: Word16

iBlocks :: Word32

iFlags :: Word32

iOsd1 :: Word32

iBlock :: (Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32)

iGeneration :: Word32

iFileAcl :: Word32

iDirAcl :: Word32

iFaddr :: Word32

iOsd2 :: ByteString

Lenses

uid :: Lens' Inode Word16

size :: Lens' Inode Word32

osd2 :: Lens' Inode ByteString

osd1 :: Lens' Inode Word32

mtime :: Lens' Inode Word32

mode :: Lens' Inode Word16

linksCount :: Lens' Inode Word16

gid :: Lens' Inode Word16

generation :: Lens' Inode Word32

flags :: Lens' Inode Word32

fileAcl :: Lens' Inode Word32

faddr :: Lens' Inode Word32

dtime :: Lens' Inode Word32

dirAcl :: Lens' Inode Word32

ctime :: Lens' Inode Word32

blocks :: Lens' Inode Word32

block :: Lens' Inode (Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32, Word32)

atime :: Lens' Inode Word32

Parsers

readInode :: Get Inode

readInodeTable :: Int -> Get [Inode]

Directory

data Directory

Constructors

Directory

Fields

dInode :: Word32

dRecLen :: Word16

dNameLen :: Word8

dFileType :: Word8

dName :: ByteString

Lenses

recLen :: Lens' Directory Word16

nameLen :: Lens' Directory Word8

name :: Lens' Directory ByteString

inode :: Lens' Directory Word32

fileType :: Lens' Directory Word8

Parsers

readDirectory :: Get Directory

Utility functions

blockGroupCount :: Superblock -> Int

Get the number of block groups within the file system.

byteFromBlock :: Superblock -> Int -> Int

Given a Superblock, and the block number, get the byte number at which it starts.

Silly test/helper functions

getSuperblock :: String -> IO Superblock

Open an ext2 filesystem and parse out the Superblock.

getBGDT :: String -> IO BlockGroupDescriptorTable

Open an ext2 filesystem and parse out the BlockGroupDescriptorTable.

getAllTables :: String -> IO (Superblock, ExtendedSuperblock, BlockGroupDescriptorTable)

Parses all tables out.